refactor(layout): 글로벌 네비게이션 바 폴더 구조 변경
