+++ app/_components/layout/HeaderCenterSection/index.tsx
@@ -1,0 +1,9 @@
+import SearchBar from './SearchBar';
+
+export default function HeaderCenterSection() {
+  return (
+    <div className="flex flex-1 justify-center">
+      <SearchBar />
+    </div>
+  );
+}
