+++ app/_components/layout/HeaderLeftSection/CategoryDropdown.tsx
